pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "devlinker"
version = "0.1.0"
description = "Multi-channel bridge for external AI coding agents such as Codex CLI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "DevLinker" }
]
dependencies = [
  "discord.py>=2.5,<3.0",
  "loguru>=0.7,<1.0",
  "pydantic>=2.7,<3.0",
  "python-dotenv>=1.0,<2.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
]
web = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
]

[project.scripts]
devlinker = "devlinker.app:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
